# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ice Truck Tracking API",
    "version": "2.0.0",
    "description": "Professional Ice Truck Tracking System API — PostgreSQL + TimescaleDB, Redis, MQTT"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Truck": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "plate_number": { "type": "string" },
          "model": { "type": "string" },
          "capacity": { "type": "number" },
          "status": { "type": "string", "enum": ["active", "inactive", "maintenance"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Shop": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "address": { "type": "string" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "TrackingRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "truck_id": { "type": "integer" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "speed": { "type": "number" },
          "heading": { "type": "number" },
          "recorded_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/trucks": {
      "get": {
        "tags": ["Trucks"],
        "summary": "Get all trucks",
        "responses": {
          "200": {
            "description": "List of trucks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "results": { "type": "integer" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Truck" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Trucks"],
        "summary": "Create a truck (requires trucks:create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Truck" }
            }
          }
        },
        "responses": {
          "201": { "description": "Truck created" },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
    "/trucks/{id}": {
      "get": {
        "tags": ["Trucks"],
        "summary": "Get truck by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Truck details" },
          "404": { "description": "Truck not found" }
        }
      }
    },
    "/shops": {
      "get": {
        "tags": ["Shops"],
        "summary": "Get all shops",
        "responses": {
          "200": {
            "description": "List of shops",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "results": { "type": "integer" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Shop" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Shops"],
        "summary": "Create a shop (requires shops:create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Shop" }
            }
          }
        },
        "responses": {
          "201": { "description": "Shop created" },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
    "/shops/{id}": {
      "get": {
        "tags": ["Shops"],
        "summary": "Get shop by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Shop details" },
          "404": { "description": "Shop not found" }
        }
      }
    },
...
    "/tracking/trucks/{id}": {
      "get": {
        "tags": ["Tracking"],
        "summary": "Get latest position for a truck (requires tracking:read)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Truck latest position" },
          "404": { "description": "Truck position not found" }
        }
      }
    },
    "/tracking/history/{truckId}": {
      "get": {
        "tags": ["Tracking"],
        "summary": "Get tracking history for a truck (requires tracking:read)",
        "parameters": [
          { "name": "truckId", "in": "path", "required": true, "schema": { "type": "integer" } },
          {
            "name": "start",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "Start of time range"
          },
          {
            "name": "end",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "End of time range"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "Max records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Tracking history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TrackingRecord" }
                    },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tracking/location": {
      "post": {
        "tags": ["Tracking"],
        "summary": "Submit a location update (requires tracking:create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["truckId", "latitude", "longitude"],
                "properties": {
                  "truckId": { "type": "integer" },
                  "latitude": { "type": "number" },
                  "longitude": { "type": "number" },
                  "speed": { "type": "number" },
                  "heading": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Location recorded" },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
    "/tracking/bulk": {
      "post": {
        "tags": ["Tracking"],
        "summary": "Bulk location update (requires tracking:create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["data"],
                "properties": {
                  "data": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "truckId": { "type": "integer" },
                        "latitude": { "type": "number" },
                        "longitude": { "type": "number" },
                        "speed": { "type": "number" },
                        "heading": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Locations recorded" },
          "400": { "description": "data must be an array of locations" }
        }
      }
    },
...
  }
}